#4647 - extracted class PropertyNameParser
